Move nanosec delay to util header to allow reuse.
